% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assignment 2.R
\name{lat_long_conv}
\alias{lat_long_conv}
\title{Latitude Longitude Conversion}
\usage{
lat_long_conv(
  input_latitude_column,
  input_longitude_column,
  flip_columns = TRUE
)
}
\arguments{
\item{input_latitude_column}{the column data of the latitude coordinates from the given data.frame}

\item{input_longitude_column}{the column data of the longitude coordinates from the given data.frame}

\item{flip_column}{If true, will provide Longitude, Latitude output. If false, will provide Latitude, Longitude output.}
}
\description{
Converts DDM coordinates to DD coordinates
}
\examples{
lat_long_conv 
}
\keyword{coordinates}
\keyword{latitude}
\keyword{longitude}
